Refactor stories
refs #31
