+++ components/form-inputs/FormInput.tsx
@@ -4,7 +4,7 @@ import TextOptional from "../text-optional/TextOptional";
 
 type FormInputType = "text" | "date" | "number";
 
-export interface FormInputProps {
+interface FormInputProps {
     /**
      * Input type. Default is 'text'
      */
