+++ components/header/Header.tsx
@@ -2,7 +2,7 @@ import { headerFont } from "@/app/fonts";
 import { ThemeColorPicker } from "@/theme/color";
 import PropsTheme from "@/theme/props";
 
-export interface HeaderProps extends PropsTheme {
+interface HeaderProps extends PropsTheme {
     /**
      * Text to display
      */
