+++ components/text-optional/TextOptional.tsx
@@ -1,4 +1,4 @@
-export interface TextOptionalProps {
+interface TextOptionalProps {
     /**
      * Is input optional. Default is 'false'
      */
@@ -9,6 +9,4 @@ export default function TextOptional(props: TextOptionalProps) {
     if (props.optional) {
         return <span className="text-gray-500">Optional</span>
     }
-
-    return null;
 }
