+++ stories/HeaderLogo.stories.tsx
@@ -1,7 +1,7 @@
 import HeaderLogo from "@/components/header-logo/HeaderLogo";
 import type { Meta, StoryObj } from "@storybook/react";
 
-const meta: Meta = {
+const meta: Meta<typeof HeaderLogo> = {
   title: "Header Logo",
   component: HeaderLogo,
   tags: ["autodocs"],
@@ -10,4 +10,8 @@ const meta: Meta = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Example: Story = {};
+export const Example: Story = {
+  args: {
+    theme: "dark"
+  }
+};
